requirements.txt
pandas==2.2.3
llama-index==0.12.41
llama-index-agent-openai==0.4.9
llama-index-cli==0.4.3
llama-index-core==0.12.41
llama-index-embeddings-huggingface==0.5.4
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.4
llama-index-llms-openai==0.4.4
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
PyYAML==6.0.2
colorama==0.4.6
python-dotenv==1.1.0
streamlit==1.45.1
sentence-transformers==4.1.0
transformers==4.52.4
torch==2.7.1
accelerate==1.7.0

# PostgreSQL and Database Dependencies
psycopg2-binary==2.9.10
llama-index-vector-stores-postgres==0.5.3
chardet==5.2.0
asyncpg==0.30.0
pgvector==0.4.1
sqlalchemy==2.0.36

# Additional Testing Dependencies
tqdm==4.66.4
numpy==1.26.4
regex==2024.9.11
